refactor(indicateurs): extrait useGridDragHandlers de GridFrame
GridFrame noyait ~90 lignes d'orchestration DnD (sensors, collision,
onDragEnd, describeDragId, annonces a11y) devant son JSX. Extrait dans
useGridDragHandlers ; GridFrame redevient une reassemblage (DndContext +
GridHead + GridBody).
